refactor: move user profile file
