+++ src/components/layout/Nav.jsx
@@ -7,7 +7,7 @@ import Sidebar from './drawer/Drawer'
 import FilterMenu from './dialogs/filters/FilterMenu'
 import UserOptions from './dialogs/UserOptions'
 import Tutorial from './dialogs/tutorial/Tutorial'
-import UserProfile from './dialogs/UserProfile'
+import UserProfile from './dialogs/profile'
 import Search from './dialogs/Search'
 import MessageOfTheDay from './dialogs/Motd'
 import DonorPage from './dialogs/DonorPage'
