+++ src/components/layout/dialogs/profile/index.jsx
@@ -5,15 +5,15 @@ import { useTranslation } from 'react-i18next'
 import { useLayoutStore, useStatic } from '@hooks/useStore'
 import Utility from '@services/Utility'
 
-import Header from '../general/Header'
-import Footer from '../general/Footer'
-import TabPanel from '../general/TabPanel'
-import { DialogWrapper } from './DialogWrapper'
-import { UserBackups } from './profile/Backups'
-import { UserPermissions } from './profile/Permissions'
-import { UserGymBadges } from './profile/GymBadges'
-import { LinkAccounts } from './profile/LinkAccounts'
-import { ExtraUserFields } from './profile/ExtraFields'
+import Header from '../../general/Header'
+import Footer from '../../general/Footer'
+import TabPanel from '../../general/TabPanel'
+import { DialogWrapper } from '../DialogWrapper'
+import { UserBackups } from './Backups'
+import { UserPermissions } from './Permissions'
+import { UserGymBadges } from './GymBadges'
+import { LinkAccounts } from './LinkAccounts'
+import { ExtraUserFields } from './ExtraFields'
 
 export default function UserProfile() {
   Utility.analytics('/user-profile')
